Refactor and add temporal CharacterBox
CharacterBox from 1ccab00ff424c2b053ab4175159d8d22d9637466

Plus refactor other branches' code to a separated file
